requirements.txt
tqdm
scikit-learn
pillow
PyYAML
pandas
torchvision
torch
timm>=0.9.10
huggingface-hub>=0.23.0
open_clip_torch
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
pandas>=1.5.0
Pillow>=9.0.0
tqdm>=4.65.0
joblib>=1.2.0